修改：unlink本地套接字时需要判断是否是服务器，客户端没有权限进行unlink，只能使用。主进程创建线程调用函数执行任务之后，线程会被销毁从而调用析构函数unlink本地套接字，解决了由客户端unlink套接字带来的服务器无法继续工作的bug

+++ ServerLinux/ServerLinux/Socket.cpp
@@ -103,7 +103,10 @@ int CSocketBase::Close()
 	m_status = 3;
 	if (m_socket != -1)
 	{
-		unlink(m_param.ip);
+		if (m_param.attr & SOCK_ISSERVER)
+		{
+			unlink(m_param.ip);
+		}
 		int fd = m_socket;
 		m_socket = -1;
 		close(fd);
